Make the async queue a queue
Right now all promises are started immediately, so the writes can happen
in the wrong order. This enforces the order of the writes by making each
item in the queue a function which we only execute when it reaches the
front of the queue. We also modify the default reducer to make a copy of
the state, so each saveState call will have a snapshot of the state
instead of the global object that can be modified.