Allow skipping intermediate saveState calls with asyncStorage
If the state is modified many times in a row, we will queue up a
saveState call for each change. This causes a write to storage for each
change. For many saveState implementations (including the default one)
this does nothing, since the state is merely serialized and written to
a storage key. In these cases, it is fine to only perform the last write
in the queue, skipping all intermediate writes.

This PR adds a new option `skipIntermediateSaves`.